requirements.txt
Flask==3.0.0
Flask-Limiter
Flask-Mail==0.9.1
tensorflow==2.15.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.2
nltk==3.8.1
deep-translator==1.11.4
requests==2.31.0
gunicorn==21.2.0
pickle-mixin==1.0.2
flask_sqlalchemy 
mysql-connector-python